Separated Diffusion into P2P, NonP2P and Common
Diffusion.hs now offers the needed API to ouroboros-consensus in a way
that it is possible to pick between P2P and NonP2P

- P2P is the p2p-master branch Diffusion module
- NonP2P is the master branch Diffusion module
- Common contains data types that are common to both versions
- Diffusion.hs indexes both P2P and NonP2P modules
- Uniforms DiffusionTracers to accomodate p2p types switching
- Uniforms DiffusionApplication to accomodate p2p types switching
- Uniforms DiffusionArguments to accomodate p2p types switching
- Renamed P2P and NonP2P DTracers, DApps, DArgs to
  DApplicationsExtra and DArgumentsExtra (these types fill the p2p
  polymorphic types in DiffusionApplication, DiffusionArguments and
  DiffusionTracers
- General long lines cleanup

Exported ErrorPolicy module
