+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -134,6 +134,7 @@ library
                        Ouroboros.Consensus.Node
                        Ouroboros.Consensus.Node.DbLock
                        Ouroboros.Consensus.Node.DbMarker
+                       Ouroboros.Consensus.Node.ErrorPolicy
                        Ouroboros.Consensus.Node.RethrowPolicy
                        Ouroboros.Consensus.Node.Exit
                        Ouroboros.Consensus.NodeId
