+++ ouroboros-network/ouroboros-network.cabal
@@ -128,7 +128,10 @@ library
                        Ouroboros.Network.TxSubmission.Inbound
                        Ouroboros.Network.TxSubmission.Mempool.Reader
                        Ouroboros.Network.TxSubmission.Outbound
-  other-modules:       Ouroboros.Network.PeerSelection.Governor.ActivePeers
+  other-modules:       Ouroboros.Network.Diffusion.Common
+                       Ouroboros.Network.Diffusion.P2P
+                       Ouroboros.Network.Diffusion.NonP2P
+                       Ouroboros.Network.PeerSelection.Governor.ActivePeers
                        Ouroboros.Network.PeerSelection.Governor.EstablishedPeers
                        Ouroboros.Network.PeerSelection.Governor.KnownPeers
                        Ouroboros.Network.PeerSelection.Governor.Monitor
